\name{ofc}
\docType{data}
\alias{ofc}
\usage{data(ofc)}
\title{OFC data}
\description{
  Simulated head size data of new borns.
}
\format{A data frame with 1000 observations on 2 variables.
  \tabular{lll}{
    \code{x0}  \tab numeric  \tab Original OFC values \cr
    \code{x}   \tab numeric  \tab OFC values rounded to centimeters \cr
  }
}
\references{
Wang, CSDA and JSS papers.
}
\keyword{datasets}
